go/{genesis,staking}: move sanity checks to a separate file
